most_seen_vertex_sssp_pregel: Given a graph and src_id(landmark) calculates number of times each vertex has been seen in the shortest paths. Doesn't return anything right now.
ssc_graphx: Given a graph and #Iterations returns strongly connected components for the graph.

cc_graphx: Given a graph cc is 'calculated' and number of subgraphs and subGraph size is calculated.
Doesn't return anything right now.

subgraphs_from_connected_components: Given a graph and Iterable with VertexId (contains all ids for the subgraph) creates a Graph and returns it.

subgraphs_from_connected_components: Given a graph and Array of subgraphs, creates all the graphs and return as Array[Graph]